+ Define

+++ Cmd.java
@@ -3,5 +3,5 @@ import java.util.Stack;
 //Created by Pavel A. on 07.03.15.
 
 public interface Cmd {
-    public Object exec(Stack<Integer> st, String[] parts);
+    public Object exec(Stack<Double> st, String[] parts);
 }
