+++ Print.java
@@ -5,7 +5,7 @@ import java.util.Stack;
 public class Print implements Cmd {
 
     @Override
-    public Object exec(Stack<Integer> st, String[] parts) {
+    public Object exec(Stack<Double> st, String[] parts) {
         return "stack: " + st;
     }
 }
